==> notebooks/vacancy_over_time/1.0 Filter data.ipynb <==
# Filter data used in vacancy model based on TCI parcel numbers

Parcels surveyed in Summer 2015 so all data pulled should come from before that. Goal of script/notebook is to filter datasets by the parcel numbers in the TCI survey, although we will filter again based on existence of structure.

import pandas as pd
import csv
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
import datetime as dt
%matplotlib inline

path = '/'.join(os.getcwd().split('/')[:-2])
luc = '/'.join(os.getcwd().split('/')[:-3])
print(path)

tci = pd.read_csv(path+'/data/model_data/tci_1_0.csv', parse_dates=['Date'], dtype={'Ward':object,'PIN':str})
ppns = set(tci[(tci.USE_CLASS=='R') & (tci.vacant>-1)].parcel)

# tci = pd.read_excel(path+'/data/inspection_data/Cleveland_Final_Results_Table_FOR_DISTRIBUTION_20151111.xlsx', encoding="ISO-8859-1") 

# def get_vacant(x):
#     if x == 'Occupied Structure':
#         return 0
#     elif x == 'Vacant Structure':
#         return 1
#     else: 
#         return -1
    
# tci['vacant'] = tci['Survey Category'].apply(get_vacant)
# tci['parcel'] = tci.PIN.apply(lambda x: x[0:3]+'-'+x[3:5]+'-'+x[5:])

# # tci[(tci.USE_CLASS=='R') & (tci.vacant>-1)].to_csv(path+'/data/model_data/tci_1_0.csv', index=False)

## Property characteristics

df = pd.read_csv(luc+'/demo.csv', parse_dates=['drdate'])

# commented code only needs to be run once

# main14 = main[main.taxyr=='2014'].groupby('parcel').first().reset_index()
# main13 = main[main.taxyr=='2013'].groupby('parcel').first().reset_index()
# main12 = main[main.taxyr=='2012'].groupby('parcel').first().reset_index()
# main11 = main[main.taxyr=='2011'].groupby('parcel').first().reset_index()
# main10 = main[main.taxyr=='2010'].groupby('parcel').first().reset_index()
# main09 = main[main.taxyr=='2009'].groupby('parcel').first().reset_index()
# main08 = main[main.taxyr=='2008'].groupby('parcel').first().reset_index()
# main07 = main[main.taxyr=='2007'].groupby('parcel').first().reset_index()

# main14.to_csv(path+'/data/clean_data/main_prop_2014.csv')
# main13.to_csv(path+'/data/clean_data/main_prop_2013.csv')
# main12.to_csv(path+'/data/clean_data/main_prop_2012.csv')
# main11.to_csv(path+'/data/clean_data/main_prop_2011.csv')
# main10.to_csv(path+'/data/clean_data/main_prop_2010.csv')
# main09.to_csv(path+'/data/clean_data/main_prop_2009.csv')
# main08.to_csv(path+'/data/clean_data/main_prop_2008.csv')
# main07.to_csv(path+'/data/clean_data/main_prop_2007.csv')


main14 = pd.read_csv(path+'/data/clean_data/main_prop_2014.csv')
main13 = pd.read_csv(path+'/data/clean_data/main_prop_2013.csv')
main12 = pd.read_csv(path+'/data/clean_data/main_prop_2012.csv')
main11 = pd.read_csv(path+'/data/clean_data/main_prop_2011.csv')
main10 = pd.read_csv(path+'/data/clean_data/main_prop_2010.csv')
main09 = pd.read_csv(path+'/data/clean_data/main_prop_2009.csv')
main08 = pd.read_csv(path+'/data/clean_data/main_prop_2008.csv')
main07 = pd.read_csv(path+'/data/clean_data/main_prop_2007.csv')

# main_all_years = main[main.taxyr.isin(['2007','2008','2009','2010','2011','2012','2013','2014','2015'])]
main_all_years = main14.append(main13).append(main12).append(main11).append(main10).append(main09).append(main08).append(main07)
main_all_years[['parcel','taxyr','pclass','propsize','front','depth','totbldgs','condition',\
        'totusabl','yrbuilt','style','tmktval','mktland','mktbldg','ownerocc','LATITUDE','LONGITUDE']]\
        .to_csv(path+'/data/clean_data/main_prop_all_years.csv', index=False)

## Residential characteristics 
Filename: ```res2014.csv```

# first part only needs to be run once

res2014 = pd.read_csv(luc+'/reschar/res2014.csv')
res2013 = pd.read_csv(luc+'/reschar/res2013.csv')
res2012 = pd.read_csv(luc+'/reschar/res2012.csv')
res2011 = pd.read_csv(luc+'/reschar/res2011.csv')
res2010 = pd.read_csv(luc+'/reschar/res2010.csv')
res2009 = pd.read_csv(luc+'/reschar/res2009.csv')
res2008 = pd.read_csv(luc+'/reschar/res2008.csv')
res2007 = pd.read_csv(luc+'/reschar/res2007.csv')
# res = res[res.parcel.isin(ppns)]
# res.to_csv(path+'/data/clean_data/res.csv', index=False)

# res = pd.read_csv(path+'/data/clean_data/res.csv')

res2014['year'] = 2014
res2013['year'] = 2013
res2012['year'] = 2012
res2011['year'] = 2011
res2010['year'] = 2010
res2009['year'] = 2009
res2008['year'] = 2008
res2007['year'] = 2007

reschar = res2014.loc[res2014.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']]\
            .append(res2013.loc[res2013.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']])\
            .append(res2012.loc[res2012.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']])\
            .append(res2011.loc[res2011.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']])\
            .append(res2010.loc[res2010.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']])\
            .append(res2009.loc[res2009.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']])\
            .append(res2008.loc[res2008.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']])\
            .append(res2007.loc[res2007.parcel.isin(ppns),['parcel','cqual','occup','rnumstor','year']])
                                

reschar.to_csv(path+'/data/clean_data/res.csv', index=False)

## Tax bill
Filename: ```dec14_tci.csv```

##only needs to be run once to clean taxbill data

infile = path+'/data/original_data/taxbill/may15.csv'
outfile = path+'/data/clean_data/taxbill_may15.csv'

cols = ['PROPERTY_NUMBER','TOTAL_NET_DELQ_BALANCE','TAX_ASSESSED_LAND','TAX_MARKET_LAND',\
        'LENDER_PROCESS_TYPE','GRAND_TOTAL_BALANCE','GRAND_TOTAL_PAID','GRAND_TOTAL_OWED']

with open(infile, 'r') as fin, open(outfile, 'w') as fout:
    write_to = csv.writer(fout, lineterminator='\n')
    header = next(csv.reader(fin))
    write_to.writerow(header)
    for row in csv.reader(fin):
        if row[5] in ppns:
            write_to.writerow(row)

months = {'jan':1,'feb':2,'mar':3,'apr':4,'may':5,'jun':6,'jul':7,'aug':8,'sep':9,'oct':10,'nov':11,'dec':12}

files = ['dec14.csv', 'sep14.csv','jun14.csv','feb14.csv','nov13.csv','aug13.csv','apr13.csv',\
         'feb13.csv','dec12.csv','aug12.csv','jun12.csv','feb12.csv','sep11.csv','may11.csv',\
         'feb11.csv','nov10.csv','oct10.csv','jul10.csv','apr10.csv','oct09.csv','dec15.csv',\
         'oct15.csv','jul15.csv','may15.csv','feb15.csv']
i = 1
for file in files:
    if i == 1:
        tb = pd.read_csv(path+'/data/clean_data/taxbill/'+file, index_col=[0])
    else:
        tb2 = pd.read_csv(path+'/data/clean_data/taxbill/'+file, index_col=[0])
        tb = tb.append(tb2)
    i+=1
tb.to_csv(path+'/data/clean_data/taxbill/tb_all.csv',index=False)
#     infile = path+'/data/original_data/taxbill/may15.csv'
#     outfile = path+'/data/clean_data/taxbill_may15.csv'
#     tb = tb[cols]
#     tb = tb[tb.PROPERTY_NUMBER.isin(ppns)]
#     tb['date'] = dt.date(int('20'+file[3:5]), months[file[0:3]],1)
#     tb.to_csv(path+'/data/clean_data/taxbill/'+file)
    
#     break
#     with open(infile, 'r') as fin, open(outfile, 'w') as fout:
#         write_to = csv.writer(fout, lineterminator='\n')
#         header = next(csv.reader(fin))
#         write_to.writerow(header)
#         for row in csv.reader(fin):
#             if row[5] in ppns:
#                 write_to.writerow(row)

# ##only needs to be run once to clean taxbill data

# infile = path+'/data/original_data/taxbill/sep14.csv'
# outfile = path+'/data/clean_data/taxbill_sep14.csv'

# with open(infile, 'r') as fin, open(outfile, 'w') as fout:
#     write_to = csv.writer(fout, lineterminator='\n')
#     header = next(csv.reader(fin))
#     write_to.writerow(header)
#     for row in csv.reader(fin):
#         if row[5] in ppns:
#             write_to.writerow(row)

## County land bank
Filename: ```count_land_bank.csv```

lb = pd.read_csv(path+'/data/original_data/count_land_bank.csv', parse_dates=[3,4])

lb = lb[lb['parcel'].isin(ppns)]
# lb = lb[lb['acq_dt']<np.datetime64('2015-06-01')]

lb.to_csv(path+'/data/clean_data/county_lb.csv', index=False)

## Foreclosure filings

fc = pd.read_csv(path+'/data/original_data/foreclosure_filings2006_beyond.csv', parse_dates = [2])

fc = fc[fc['parcel'].isin(ppns)]
# fc = fc[fc['filedate']<np.datetime64('2015-06-01')]

fc.to_csv(path+'/data/clean_data/foreclosure_filings2.csv',index=False)

# fc2 = pd.read_csv(path+'/data/original_data/foreclosure_filings2006_dec2014.csv', parse_dates = [2])

## Sheriff auction
Filename: ```shf_aution_mar2000_dec2014.csv```

sa = pd.read_csv(path+'/data/original_data/shf_aution_mar2000_dec2014.csv', parse_dates=[2], encoding="ISO-8859-1")

sa = sa[sa.parcel.isin(ppns)]
# sa = sa[sa.salesdt<np.datetime64('2015-06-01')]

sa.to_csv(path+'/data/clean_data/sheriff_auction.csv', index=False)

## Transfers
Filename: ```transfers2000_2014.csv```

infile = path+'/data/original_data/transfers2000_beyond.csv'
outfile = path+'/data/clean_data/transfers.csv'

with open(infile, 'r') as fin, open(outfile, 'w') as fout:
    write_to = csv.writer(fout, lineterminator='\n')
    header = next(csv.reader(fin))
    write_to.writerow(header)
    for row in csv.reader(fin):
        if row[5] in ppns:
            write_to.writerow(row)

# tf = pd.read_csv(path+'/data/clean_data/transfers.csv',parse_dates=[8], dtype='str')
# # tf = tf[tf.mdate<np.datetime64('2015-06-01')]

# tf.to_csv(path+'/data/clean_data/transfers.csv',index=False)

## Armslength sales
Filename: ```armslengthsales2006_2014.csv```

# al = pd.read_csv(path+'/data/original_data/armslengthsales2006_beyond.csv', dtype=str)

al = pd.read_csv(path+'/data/original_data/armslengthsales2006_beyond.csv', dtype=str)
al = al[al.PROPERTY_NUMBER.isin(ppns)]
al.to_csv(path+'/data/clean_data/armslength.csv',index=False)

# al = pd.read_csv(path+'/data/clean_data/armslength.csv', parse_dates=['mdate'])

## Violations
Filename: ```violate_cle.csv```


infile = path+'/data/original_data/violate_cle.csv'
outfile = path+'/data/clean_data/violations.csv'

v = pd.read_csv(infile, dtype=str)
v = v[v.parcel.isin(ppns)]
# v = v[(v.v_file_date < np.datetime64('2015-06-01')) & (v.v_file_date > np.datetime64('2006-06-01'))]

# v.to_csv(path+'/data/clean_data/violations.csv',index=False)

## Complaints
Filename: ```complaint_cle.csv```

c = pd.read_csv(path+'/data/original_data/complaint_cle.csv', dtype=str)

c = c[c.parcel.isin(ppns)]
# c = c[(c.c_file_date < np.datetime64('2015-06-01')) & (c.c_file_date > np.datetime64('2006-06-01'))]

# c.to_csv(path+'/data/clean_data/complaints.csv')

## Postal data
Filenames: ```pv201302.csv, pv201304.csv, pv201308.csv, pv201312.csv, pv201402.csv```

tci.shape

p = pd.DataFrame()
for pv in os.listdir(path+'/data/original_data/postal/'):
    print(pv)
    pos = pd.read_csv(path+'/data/original_data/postal/' + pv)
    pos['date'] = dt.datetime(int(pv[2:6]), int(pv[6:8]), 1)
    p = p.append(pos)
p = p[p.PARCEL.isin(ppns)]
p.to_csv(path+'/data/clean_data/postal_vacancy.csv')
