==> notebooks/vacancy_over_time/3.0 Combine data.ipynb <==
# 3.0 Combine data

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
import matplotlib.pyplot as plt
import numpy as np
import seaborn as sns
import math
from sklearn.preprocessing import Imputer
from sklearn.base import TransformerMixin
from sklearn import cross_validation
from sklearn.grid_search import GridSearchCV
from sklearn.linear_model import LogisticRegression
from sklearn import preprocessing
from sklearn.cross_validation import cross_val_score
import os

def sound():
    os.system('afplay /System/Library/Sounds/Sosumi.aiff')

%matplotlib inline
sns.set_context('notebook')

path = '/'.join(os.getcwd().split('/')[:-2])
print(path)

cols = ['Date','vacant','Survey Category','Survey Grade','SPA_NAME']

rng = pd.date_range('10/1/2009', periods=23, freq='3MS')

for date in rng:
    print(date)
    df0 = pd.read_csv(path+'/data/model_data/tci_2_0_'+str(date.year)+'_'+str(date.month)+'.csv', \
                      dtype={'Parcel ID':str}, index_col='parcel')
    df1 = pd.read_csv(path+'/data/model_data/tci_2_1_'+str(date.year)+'_'+str(date.month)+'.csv', \
                      dtype={'Parcel ID':str}, index_col='parcel')
    df2 = pd.read_csv(path+'/data/model_data/tci_2_2_'+str(date.year)+'_'+str(date.month)+'.csv', \
                      dtype={'Parcel ID':str}, index_col='parcel')
    df3 = pd.read_csv(path+'/data/model_data/tci_2_3_'+str(date.year)+'_'+str(date.month)+'.csv', \
                      dtype={'Parcel ID':str}, index_col='parcel')
    df = pd.merge(df0, df1[[col for col in df1.columns if col not in cols]], how='inner', left_index=True, right_index=True)
    df = pd.merge(df,  df2[[col for col in df2.columns if col not in cols]], how='inner', left_index=True, right_index=True)
    df = pd.merge(df,  df3[[col for col in df3.columns if col not in cols]], how='inner', left_index=True, right_index=True)
#     break
    df['price_per_sqft'] = df['tmktval']/df['totusabl']
#     df = df.set_index('parcel')
    # df['missing_values'] = df.apply(lambda x: len(x[x.isnull()]),axis=1)
    to_remove = ['Parcel ID','Date','Survey Category','Survey Grade','date'\
             'condition','SPA_NAME','NAME10','style_filtered','rextwall','LENDER_PROCESS_TYPE']
    
    features = list(df.columns.values)
    for i in to_remove:
        if i in features:
            features.remove(i)

    df[features].to_csv(path+'/data/model_data/tci_3_0_'+str(date.year)+'_'+str(date.month)+'.csv', \
                      index=True)

df.info()
